requirements.txt
mplayer.py (and mplayer)
